Relocate `HealthResponse` model class definition to `schemas.py` module
